chore!: change Hook Config field from map to *HookConfig
